refactor(cli/audio): split decoder_thread submodule into file
